Integrate mermaid hurt animations
