+++ Assets/Animations/Mermaid/Hurt.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: bcdca677456ef5f409b943041148b2b6
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
