+++ Assets/Animations/Mermaid/Hurt/Mermaid Damage Taken Left.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 40c6edc296320794ea1d725aa6519733
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
