+++ Assets/Animations/Mermaid/Hurt/Mermaid Damage Taken Right.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 02eaa2774687f0c41aa364b156b348c4
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
