+++ Assets/Animations/Mermaid/Hurt/Mermaid Die From Left.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2f9d94ad13771274a8af34acb5e6fe52
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
