+++ Assets/Animations/Mermaid/Hurt/Mermaid Die From Right.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 600ab10a996dab440b916b99cea0d02d
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
