+++ Assets/Animations/Mermaid/Hurt/Mermaid Fall From Left.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: fb8168e2c60ff6b49a4c8883673a3cf9
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
