+++ Assets/Animations/Mermaid/Hurt/Mermaid Fall From Right.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 619b0dcf1581e8b438d80437975f88f3
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
